pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pypii"
version = "0.1.0"
description = "PII Detection Tool"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "python-magic>=0.4.27",
]

[project.scripts]
pypii = "pypii.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["pypii"]